minor fixes ^^

+++ app/src/main/java/nsu/krpo/academads/data/network/models/BookingRequest.kt
@@ -3,12 +3,12 @@ package nsu.krpo.academads.data.network.models
 import com.squareup.moshi.Json
 import java.math.BigDecimal
 
-data class BookingRequest(
+data class BookingRequest       (
         @Json(name = "ads_id")
         val adsId: Long,
 
         @Json(name = "claimant")
-        val claimant: BigDecimal,
+        val claimant: Long,
 
         @Json(name = "dateUntil")
         val dateUntil: String
